+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * refactor: move `refreshCount` to its own interface
 * refactor: rename `getReference` to `reference`
 * refactor: rename the term 'basic' to 'minimal'
+* test: refactor tests
 
 ## 0.3.0